Deduplicate code for PlayFirst
