+++ FredBoat/src/main/java/fredboat/commandmeta/CommandInitializer.kt
@@ -240,9 +240,9 @@ class CommandInitializer(cacheMetrics: CacheMetricsCollector, weather: Weather, 
         musicModule.registerCommand(PlayCommand(playerLimiter, trackSearcher, videoSelectionCache,
                 listOf(SearchProvider.YOUTUBE, SearchProvider.SOUNDCLOUD),
                 "playnext", "playtop", "pn", isPriority = true))
-        musicModule.registerCommand(PlayFirstCommand(playerLimiter, trackSearcher, videoSelectionCache,
+        musicModule.registerCommand(PlayCommand(playerLimiter, trackSearcher, videoSelectionCache,
                 Arrays.asList(SearchProvider.YOUTUBE, SearchProvider.SOUNDCLOUD),
-                PLAY_COMM_NAME, "pf", "lucky"))
+                PLAY_COMM_NAME, "playfirst", "pf", "lucky", skipSearch = true))
         musicModule.registerCommand(PlaySplitCommand(playerLimiter, "split"))
         musicModule.registerCommand(RepeatCommand("repeat", "rep", "loop"))
         musicModule.registerCommand(ReshuffleCommand("reshuffle", "resh"))
