+++ FredBoat/src/main/resources/lang/en_US.properties
@@ -226,6 +226,7 @@ helpLeaveCommand=Make the bot leave the current voice channel.
 helpPauseCommand=Pause the player.
 helpPlayCommand=Play music from the given URL or search for a track. For a full list of sources please visit {0}
 helpPlayNextCommand=Play music from the given URL or search for a track. Tracks are added to the top of the queue. For a full list of sources please visit {0}
+helpPlayFirstCommand=Play music from the given URL or search for a track. This command automatically selects the FIRST result from a search. For a full list of sources please visit {0}
 helpPlaySplitCommand=Split a YouTube video into a tracklist provided in it\'s description.
 helpRepeatCommand=Toggle between repeat modes.
 helpReshuffleCommand=Reshuffle the current queue.
